Autogenerate identities from plugins (#2134)
This integrates the identity proposal system defined in #2128, building
on the plugin-specific createIdentities methods in #2131, #2132, #2133.
We now generate user identities during the cli/graph command.

I also fixed that the discord identities should have `discord` as the
plugin name, and added a mock identity generator for the initiatives
plugin.

Test plan: Manual testing; I used the new system on the SC official
instance and verified that identities were generated as expected. Also,
if you run `yarn start` you'll see that there's now an identity for each
userlike from each plugin.

Note: Bots are now included in the main score page, which is a slight
regression from the previous system, but worth it. We can filter out
bots again later, in a more principled + general purpose way.

Big progress on #2109.