Create LedgerContext to manage ledger state in UI
Passing down the ledger/setLedger props into all the components was buggy and verbose. This PR refactors it to use React Context instead. Any component that
needs to consume the ledger can do so with the "useLedger" hook, which returns the ledger object as well as an "updateLedger" function that triggers a React
state update.

Test Plan: Run the frontend with yarn start and ensure all the ledger functionality is intact and the state / UI is being updated properly