Fix "Payload too large" error when saving Ledger to disk from UI
The default limit for request bodies in express is 10kb, which was too low for ledger files that had a decent amount of data.
This updates the limit to 50mb for both the `sourcecred serve` command and when using the webpack dev server with `yarn start`.

Test Plan: Load a large ledger into the admin UI, make some changes, click "Save ledger to disk", and ensure the ledger saves
to disk without any errors.