Support merging identities in the ledger admin (#2137)
As of #2134, we create identities for every user. Thus, we no longer
want to add aliases to identities, but instead we merge them together
(this functionality from #2117). This commit updates the frontend so we
can merge identities.

Thanks to @hammadj for major help letting this land in a bug free
fashion.

Test plan: Run `yarn start`, and merge some identities together.