implement custom currencies
Custom currencies can now be configured in an instance's
config/grain.json file. The fields `currencyName` and currencySuffix`
will be parsed by the frontend to display these values in place of the
default "Grain" and "g", respectively.

test plan: run `yarn start` and ensure that the "Grain" name and "g" are
not visible in the Menu bar, in the transfer arrow when selecting an
amount, or in the Accounts view. In each of these locations, they should
be replace with Dai. Other currencies can be tested by modifying the
sharness example instance's grain/config file