Remove references to grain in UI component filenames
Because the frontend now generically renders configured currencies, the
grain name might be confusing for anyone unfamiliar with the codebase.

Note: The h1 element in the SpecialDistribution component still contains
"Grain" because we want people to ask questions about using this for
now. Happy to remove it though

test plan: these changes should yield absolutely no functional changes
in behavior