Add a dependencies config type
This module defines configuration for the Dependencies system, a system
which allows a project to mint excess Cred for its dependencies.

To learn about the semantics of the dependencies system, read the module
docstring for core/dependenciesMintPolicy.js

At a high level, this config type allows the instance maintainer to
specify identities (usually PROJECT-type identities) to mint extra Cred
over time, as a fraction of the baseline instance Cred.

In the future, we'll likely build a UI to manage this config. However,
right now it's designed for hand-editability. Also, we really want to be
able to ship a default config that adds a SourceCred account (if one
doesn't already exist), activates it (if it was just created), and then
flows it some Cred.

With that in mind, here's how the config works:
- User makes a new config, specifying a name for the identity. The user does
  not manually write in an id.
- The config is validated against the ledger. If the config has an id, we
  verify that there's a matching identity in the ledger with that name (erroring
  if not). If the config doesn't have an id, we check if there is an identity
  in the ledger with that name. If there is, we write the id into the config.
  If there isn't, we create a new identity with the name, activate it (if told
  to do so by the config), and then write the id into the config.
- Afterwards, we save the config (which is guaranteed to have an id) back to
  disk.

You'll note that the state in the config is a mix of human generated
(choosing the name) and automatically maintained (the id). It's
something of a weird compromise, but it accomplishes the design
objective of having state that's easy for humans to write by hand, but
also tracks the vital information by identity id (which is immutable)
rather than by name (which is re-nameable).

Note that at present, when the identity in question is re-named, the
config must be manually updated to account for the rename. In the future
(when the config is automatically maintained) we'll remove this
requirement. (Likely we'll stop tracking the identities by name at all
in the config; that's an affordance to make the files generatable by
hand.)

Test plan: Unit tests added.