Implement custom currencies in UI (#2145)
* implement custom currencies

Custom currencies can now be configured in an instance's
config/grain.json file. The fields `currencyName` and currencySuffix`
will be parsed by the frontend to display these values in place of the
default "Grain" and "g", respectively.

test plan: run `yarn start` and ensure that the "Grain" name and "g" are
not visible in the Menu bar, in the transfer arrow when selecting an
amount, or in the Accounts view. In each of these locations, they should
be replace with Dai. Other currencies can be tested by modifying the
sharness example instance's grain/config file

* [remove test configs from config/grain.json]

* Remove references to grain in UI component filenames

Because the frontend now generically renders configured currencies, the
grain name might be confusing for anyone unfamiliar with the codebase.

Note: The h1 element in the SpecialDistribution component still contains
"Grain" because we want people to ask questions about using this for
now. Happy to remove it though

test plan: these changes should yield absolutely no functional changes
in behavior

* Update GrainConfig parser to include currency configs

The GrainConfig parser now supports parsing `currencyName` and
`currencySuffix` as optional entries in the grain.json file. The parser
is now also utilized to parse this file in the browser.

test plan: No breaking changes in frontend behavior. This just makes
reading the file from disk a bit safer and ensures the file is formed
correctly. Ensure currency names and suffixes update in the frontend as
the config file is updated

* implement currencyDetails type and parser

test plan: ensure creation of currencyDetails.json and utilization
of config fields "currencyName" and "currencySuffix" are reflected
in the Transfer, and Accounts views

* show up to two decimals in Accounts overview

test plan: ensure accounts with fractional amounts render out to at most
two decimal places

* address last review "nits"

- utilize nullUtil.orElse to elminiate empty-string special case
- rename "grain" route to "accounts" (nice catch!)
- mark currencyDetails as a required Prop in the Overview frontend to
statically handle unexpected errors where props aren't passing down

test plan: Empty strings should no longer fall back to the default case
when utilized in currencyDetails.json as either the name or the suffix