credView: retrieve intervals and intervalIndices
This makes three changes to the CredView API:

- We can now retrieve intervals directly from the CredView via
`view.intervalEnds()`.
- Nodes now have an `intervalIndex`, which finds the interval for which
that node's timestamp is included, assuming the timestamp is non-null
- Edges now have an `intervalIndex`, as with nodes (but never null)

This makes it a lot easier to do queries like finding the total cred by
interval or finding the top contributions in a given interval.

Test plan: Unit tests added. `yarn test`