integrate dep cred minting in compute pipeline
This commit updates the Cred computation pipeline so that it
incorporates dependency minting. We incorporate the dependency policies
in the CredResult to ensure that the front-end will be able to
faithfully re-compute the Cred using data available to the frontend.

Test plan: This is the kind of "glue-pieces-together" code where I
prefer to rely on flow rather than unit tests. I have tested this in a
downstream commit which integrates the logic in an end-to-end example.