Integrate dependency minting (#2156)
This commit integrates the dependencies system, allowing projects to
mint a fraction of their baseline Cred flow to send to their
dependencies. This is the final coupling of a long stream of work;
look at the PRs linked to #1941 for context.

In this commit, we wire the new functionality for reading, updating,
using the dependencies config data into the `cli/score` command.

Test plan:

I've integrated it into the test-instance. Run `yarn start`
and you'll see a `SourceCred` entry with dependency minting as its
source of cred. Run `yarn test --full` and you'll see the snapshots are
stable.

Fixes #1941.