Stop depending on userTypes
Now that all users are made into identities (cf #2134), we can hardcode
in the IDENTITY_PREFIX rather than depending on plugins to enumerate all
the user types.

I updated the credResult test case to make all of the nodes
psuedo-identity nodes, to maintain test consistency.

Test plan: `yarn test --full` passes, which means snapshots for the
example instance are unchanged, which means we're still running cred
calculation with the right set of users.