Create composable ScopedTaskReporter
ScopedTaskReporter is a higher-order task reporter for generating
opaque scopes meant to be passed into child contexts. In this case, a
scope is a log prefix indicating the parent context in which the
current task is running. This allows for reliable filtering and
searching on existing tasks by prefix. Any instance of a TaskReporter
may be used with the scopedTaskReporter.

test plan: unit tests included

Co-authored-by: William Chargin <wchargin@gmail.com>