Clean up logging
Added a fail function, which will log out the messages and use chalk to
conform to the task runner message style. Because the TaskRunner doesn't
include semantics for failures, failed tasks do not yet update runner
state and the failed entry is left dangling in the runner as before.

Utilize std implementation instead of console in all logging lines

test plan:
0. unset any discord bot tokens in your environment
1. in the example instance, run `path/to/sourcecred.js load`
2. execute this sqlite command to poison the cache db:
  sqlite3 cache/sourcecred/github/*.db "UPDATE meta SET config = 'bad'"
3. run `path/to/sourcecred.js load` again
4. Observe all failed runs logged out with chalk in their proper
  contexts.