Add Scoped Task Reporter (#2164)
* Create composable ScopedTaskReporter

ScopedTaskReporter is a higher-order task reporter for generating
opaque scopes meant to be passed into child contexts. In this case, a
scope is a log prefix indicating the parent context in which the
current task is running. This allows for reliable filtering and
searching on existing tasks by prefix. Any instance of a TaskReporter
may be used with the scopedTaskReporter.

test plan: unit tests included

* Clean up ScopedTaskReporter docstring

- add example to elucidate a situation where the reporter can be used to
generate an error state
- clean up language, spelling and punctuation

test plan: review the ScopedTaskReporter docstring and ensure it is
helpful, yet still reasonably concise

Paired with @wchargin

Co-authored-by: William Chargin <wchargin@gmail.com>