add cache-clearing step in load cli command (#2161)
* Add cache-clearing step in load cli command

This is a relatively naive implementation of a cache clearing retry for
the load command. It works as follows:
1. check all plugins to be loaded for an existing cache db and save the
state to a Map
2. For each plugin if the load fails for any reason and a cache exists,
  the cache is deleted and the load is retried. Otherwise the load fails
  as before.
3. If the single retry fails, the load will fail.

Some alternative implementations were considered but decided against for
the following reasons:
- the CLI context has no knowledge of the db file name composition
semantics, so naively checking the cache dir was implemented in favor of
looking for a specific file
- It's difficult to catalog all reasons the cache may be
invalidated across plugins, so this implementation does not parse the
error type in order to determine whether a retry is the proper next step

test plan:
1. run `path/to/sourcred.js load` to set up an initital cache
2. perform an action that would cause a sqlite db error or a
graphql error using the master cli load command
3. run `path/to/sourcred.js load` again to verify the error does not
occur. Instead, the cache db should be deleted and recreated

* Address review comments from @wchargin and @decentralion

- track tasks within plugins using a child reporter, and manually finish
all active child tasks when an error is caught
- utilize scopes to minimize  diffs and reifications
- implement warning logs so retries are explicitly printed
- rename `promise` to `loadWithPossibleRetry`

* Utilize composable ScopedTaskReporter

test plan:
0. unset any discord bot tokens in your environment
1. in the example instance, run `path/to/sourcecred.js load`
2. execute this sqlite command to poison the cache db:
  sqlite3 cache/sourcecred/github/*.db "UPDATE meta SET config = 'bad'"
3. run `path/to/sourcecred.js load` again
4. Observe the first sourcecred github load attempt fail, then reload
  successfully after deleting the cache
5. Observe that the discord plugin never attempts to reload when the
  cache db is not present

* Clean up logging

Added a fail function, which will log out the messages and use chalk to
conform to the task runner message style. Because the TaskRunner doesn't
include semantics for failures, failed tasks do not yet update runner
state and the failed entry is left dangling in the runner as before.

Utilize std implementation instead of console in all logging lines

test plan:
0. unset any discord bot tokens in your environment
1. in the example instance, run `path/to/sourcecred.js load`
2. execute this sqlite command to poison the cache db:
  sqlite3 cache/sourcecred/github/*.db "UPDATE meta SET config = 'bad'"
3. run `path/to/sourcecred.js load` again
4. Observe all failed runs logged out with chalk in their proper
  contexts.