Fix dependency minted cred amount (#2159)
This fixes a serious issue with #1941 where the dependency minted cred
was calculated as a proportion of the total instance Cred, rather than
the total participant Cred. This resulted in the dependency minted cred
being a much higher proportion of participant Cred than intended.

Since we do Grain Distributions based on participant Cred, and the UI
shows Participant Cred, this distinction matters!

I've updated the test case to reflect the correct semantics.

Test plan: Run `yarn start`, and observe that the test instance now
mints the correct amount of Cred for SourceCred (just under 5%).