Revert change to special distribution policy type (#2171)
PR #2165 renamed the "recipient" field in the special policy to
"recipientId". This is a reasonable change in isolation, but it actually
breaks existing ledgers because they already have serialized policies.
It wasn't caught in tests because the ledger snapshot tests do not
exhaustively include every policy type.

We should find a way to catch issues like this in CI. Maybe we could
enumerate every data type that is part of the serialized API, and then
have a test suite that must contain a snapshot of every serialized API
data type? Not sure. But for now, I'm just going to fix this by
reverting the field rename.

Test plan: After rebuilding with this patch, we can again load the
official SourceCred instance.