Account Overview: make table fixed height
This refactors the Account Overview page so that we now set a max height
(40 ems) on the table, and the user can scroll within the table to find
more entries. This is cleaner than simply having the table spill over
infinitely.

We might want to choose a more principled size than 40 ems (e.g. a % of
view) but I couldn't get that to work and I figured this minimal change
is still an improvement. I'd also like for us to set up scrollbar styles
that are more in line with our dark theme.

Test plan: Load with an instance which has many accounts (i.e. not the
test instance) and observe the nice scrolling behavior. Remember that
you can run `yarn start --instance PATH_TO_INSTANCE`.