Prettier: preserve quotes in object props
In the latest release of Prettier, number keys are "unquoted" by
default. This is all fine and well, but Flow reacts poorly to number
literals as keys. The most diplomatic solution seems to be setting
"quoteProps" to "preserve":
	https://prettier.io/docs/en/options.html#quote-props

However, this precipitated an interesting error in jest, where it was
unhappy about the reaction and review state props being unquoted, both
in the flow type notation and the object literal itself. I recommend
reading the PR that implemented this change for more discussion:
https://github.com/prettier/prettier/pull/8508

test plan: `yarn unit`, `yarn flow`, `yarn check-pretty`