Add parsers for TimestampMs and TimestampISO (#2226)
This commit adds parsers to the timestamp module. To accomplish this, I
wrote a new `validateTimestampISO` function, which simply checks whether
the timestamp can be converted without error, and renamed `validate` to
`validateTimestampMs` for disambiguation.

Test plan: Added tests for the `validateTimestampISO` method, although I
think it's not really necessary. Didn't test the parsers because their
implementations are trivially correct.