Track intervals and not just interval endpoints
This commit changes the cred output data pipeline to track full
`Intervals` (e.g. `{startTimeMs: 0, endTimeMs: 100}`)  rather than just
interval endpoints (e.g. `100`). The extra information makes it a lot
easier to reason about the semantic meaning of the interval, and to
choose whether we should make comparisons to the start or end of the
interval. In that sense this is a continuation of #2154, and for similar
logic.

This is a pre-refactor for fixing #2239. Without this fix, the logic
needed to fix #2239 would be somewhat opaque and brittle.

Test plan: Unit tests are updated; `yarn test` passes. Fortunately, our
time-dependent logic is generally well tested.

Since we used to just keep the interval endpoints, all the new code
referencing the intervals should exclusively reference `endTimeMs`, not
`startTimeMs`. Reviewers should check for that. :)