Fix issue in calculating Grain distributions (#2241)
This commit fixes #2239. See the bug for context. Honestly, the API for
the Grain command is really wonky and should probably be changed more,
but for now I'll settle for just fixing the bug.

Test plan: There's a unit test added for the case where the latest
interval is complete.