dependencies config: specify all deps via periods (#2242)
This removes the `startWeight` setting for the dependencies config, in
favor of making `periods` mandatory. This is a simpler API, as discussed
[here](https://github.com/sourcecred/sourcecred/pull/2228#discussion_r486663635).

It also discourages having "infinite-time" dependencies, in favor of
choosing when the dependency truly began--which is semantically better.

One interesting thing to consider: how can we have the default period
for dependency on SourceCred consistently start around the time when
someone forks the template instance? I considered setting up a GitHub
action for this, but then it would keep getting updated once people fork
the template repo. Food for thought.

Test plan: `yarn test` passes. Note this is a breaking change, but AFAIK
the only user right now is the template instance; I'll fix it when I
upgrade the instance to this version.

Note that running `yarn start` we still see dependency minting to
SourceCred after this change. (The test instance has been updated.)