Simplify CLI version command to use version from package.json
This will make it easier to keep the package version and the CLI version in sync by directly
using the version in package.json instead of maintaining a separate versioning system.

I also removed some of the extra formatting and git-state handling logic in the file since
I didnt see it being used anywhere and @decentralion mentioned that it wasn't as necessary
now that we have the NPM package as the main entrypoint for users.

Test Plan: Run `sourcecred --version` and ensure it returns "v0.7.0-beta-16", or whatever is currently
in the package.json