Mui explorer regressions and spacing #2253
In master currently, the weights explorer page has spacing issues. This is because we partially converted to use MUI grid components, but still had a number of raw html elements that don't work well with the MUI stylings. Converting these raw html elements in the Explorer and WeightConfig files to MUI components allows us to fix these spacing issues. It also improves our consistency since we are gradually switching to using MUI everywhere in our UI.

I have also removed some extra indentation in the CredRow file and changed alignment in the TableCells within the Explorer file. This improves the alignment and look of the Table on the Explorer page.

The Collapse element has been removed from CredRow. This element served no functional purpose and was causing an error in the console because it was a non-table element within the table portion of the DOM tree.

Test Plan: unit tests are passing, check to make sure UI still works as expected.