credrank: enumerate scoring addresses explicitly (#2280)
Summary:
Prior to this change, the Markov process graph took a list of node
address prefixes to determine the scoring addresses. It now takes a set
of scoring addresses directly. This is more flexible, since it enables
specifying as scoring nodes a proper subset of the nodes of a given
type.

Test Plan:
Running `sourcecred credrank` on the SourceCred cred instance outputs an
identical `output/credGraph.json` before and after this change.

wchargin-branch: credrank-enumerate-scoring-nodes