DRAFT2: credrank: add dependency cred
This is a functional version of #2282. It still needs testing.

One small test: run `credrank` inside the sc cred instance. Currently,
mints about 7x more dependency cred for SourceCred than is intended. We
should demonstrate that the mint amount is stable around our intention,
and independent of the choice of beta.

Paired with @wchargin, see #2282