Refactor Explorer to Functional Component
Split up the mega-component into three smaller functional components. Can move them to separate files,
kept them in the same one to make it easier to review changes. This will allow us to use memoized selectors
for the cred-view data, making it easier and more performant to compute derived state on the client side.

Test Plan:
Ensure the Explorer functions as it was before the refactor (changing weights, filters, etc)