Set default period if it doesn't exist in dependenciesConfig
This fixes the issue of having to keep the "startTime" value up to date in the template
instance by setting it automatically to one week from the time the dependencyId was first
created in the ledger (i.e. when the user first added the dependency). The default weight
is set to 0.05, or 5% of the total cred.

Test Plan: Create a dependencyConfig that has an empty array for periods, run sourcecred
commands to generate the instance, and ensure the dependencyConfig.json is updated with a
period startTime of one week from the current time.