flow: apply types-first codemod to `src/`
Doesn’t quite typecheck. But it’s closer.

Generated with:

```
$ yarn flow codemod annotate-exports --write --repeat src/
```

The codemod brokenly embeds an absolute path-on-disk import for one
file (`git grep /home/wchargin`). Not sure why. If I run the codemod
from a different directory, it creates *different* absolute imports.
Will just go ahead and fix this with the rest.