flow: fix and enable “types first” mode (#2304)
Summary:
“Types First” is a new architecture for Flow that is now on by default
and will be mandatory by January 2021:
<https://medium.com/flow-type/types-first-a-scalable-new-architecture-for-flow-3d8c7ba1d4eb>

The change basically amounts to “module boundaries must be explicitly
typed rather than inferred”, which is generally a good thing for
readability anyway.

Currently, we’ve opted out of types-first mode. This patch fixes all the
errors (150-ish of them) and opts us in.

This is written with the help of a codemod (~400 lines generated by
`yarn flow codemod annotate-exports --write --repeat src/`) but with
many manual changes on top (~350 lines), amounting to:

  - migrating the `config/` subtree manually, using comment syntax
  - fixing the inferred type for `Schema.unfaithful`
  - fixing an incorrectly absolute import path
  - defining a typed interface for the `api` module
  - replacing `$IMPORTED_TYPE$` placeholders with direct resolutions
  - retargeting and opacifying `"React"`
  - rewriting imports from `"filename.js"` (rather than `"filename"`)
  - enabling types-first mode in `.flowconfig`

Test Plan:
That `yarn flow` passes is the main thing; that `yarn test` passes
primarily indicates that the codemod didn’t break any import paths. (Or,
rather, that I fixed all the import paths that it broke.)