GitHub: Retry when hitting API rate limit
This modifies the GitHub fetchGithubRepo file so that when the GitHub
API rate limit is exceeded, we retry. We also modify the retry
parameters to ensure that our last retry will be after just after one
hour, which definitely gives time for us to hit the next API rate limit
reset.

We are using the same retry strategy for every error. This means that on
RATE_LIMIT_EXCEEDED, we'll retry after 1s, 2.8s, 22s, etc. These early
retries are really unlikely to work, since the API rate limit is reset
on an hourly cadence. However, the retry library doesn't allow modifying
the retry strategy in flight, so let's take this expedient hack. I don't
think the 7 extra failed API calls will bother GitHub overmuch.

Test plan: The retry logic is untested, which is reasonable as it would
be finicky to test, and doesn't get modified very much. As a test plan,
I'll try loading a large instance which is known to fail due to rate
limit exceeded, and see if it succeeds instead.

Fixes #2201