credrank: load weighted graph with custom weights
This fixes the credrank command so it properly uses the instance custom
weights. Rather than copy-paste the logic, I moved it to a helper
function, which helps declutter both score and credrank. (I intend to
keep pulling out helper functions of this form as we add identity and
dependency support.)

Test plan: Run `sourcecred credrank` and sanity check results to verify
the logic. This is unlikely to be broken in subtle or nonobvious ways.