Implement & use API for retrieving credrank scores
This commit adds functional APIs for retrieving participants and their
Cred scores from the Cred Graph.

As of #2279, we no longer embed participant (scoring) nodes in the
Markov Process Graph. However, this means the CredGraph API for
retrieving user scores was rendered non-functional.

This commit fixes this, while munging several changes together:
- We track Participants rather than scoring addresses, so that we have
descriptions available for participants (since the nodes are no longer
in the graph, we need to store their descriptions separately)
- The MarkovProcessGraph now reports all the epoch boundaries, which
makes it feasible to enumerate all the payout edges
- We expose logic for constructing a payout edge address given the epoch
time boundary and the epoch owner
- We update the CredGraph so that instead of the (non-functional)
`scoringNodes` api, it has a `participants` iterator which returns
participants along with their total Cred and their Cred over Time.

Finally, I've updated the `credrank` command so that it prints a table
of the top-10 participants and their total Cred. This allows us to
replace the "run this complicated jq command" test plan with simply
running `credrank` and observing the results. Naturally, once we have
better test infrastructure in place, we will remove this logging.

Test plan: Run `credrank` in the SourceCred cred instance, and observe:

| Description | Cred |
| --- | --- |
| github/decentralion | 5117 |
| discord/decentralion#1337 | 4927 |
| discourse/decentralion | 3807 |
| github/wchargin | 3148 |
| discord/LBS#6541 | 2773 |
| discord/s_ben#7872 | 2327 |
| discourse/LB | 1773 |
| discourse/@s_ben | 1773 |
| discord/Beanow#5887 | 1645 |
| discord/hammadj#6801 | 1393 |

Note to reviewer: This could have been split into smaller / more atomic commits, but given that the code is untested and generally in a "high-heat" early stage of simulated annealing, munging this together made sense, and made for a much easier test plan. Let me know if you have pushback!