markovProcessGraph: take intervals & refactor `new`
This modifies the MPG so that it takes the time intervals as an explicit
argument, rather than implicitly choosing intervals based on week-long
period. This will make the MPG easier to test. I also refactored the
`new` method so that it takes an object of arguments, and combined the
`seedOptions` and `fibrationOptions` into `parameters`.

This is mostly just a refactor, but we're using a slightly different
implementation for choosing weeklong intervals from a graph, such that
we now include node timestamps when picking intervals, not just edge
timestamps. This shouldn't matter in practice.

Test plan: Cred scores are unchanged.