github: retry rate limits every 15 minutes
Summary:
When a query in `fetchGithubRepo` fails because the rate limit has been
exceeded, it will now be retried after 15 minutes rather than simply
failing the load.

Fixes #2201. Thanks to #2312 for drafting an initial approach to this
problem.

Test Plan:
None; the functionality is tested in `util/retry.test.js`, and this is
just a policy change.

wchargin-branch: github-retry-rate-limits
wchargin-source: 581749826cb93b3397bcfde18bce606567b5de43