helper fn for prepping graph, ledger, and deps (#2314)
This adds a cli/common helper method for loading the graph, ledger, and
dependency policies for Cred computation. These three are intertwined in
that the dependency policies may modify the ledger, and the ledger then
modifies the graph, so it makes sense to have a single helper that
orchestrates this. I also added helpers for loading/saving the ledger.
This commit refactors the existing CLI score command to use the helper;
it will make a future change to CredRank easier as well.

Test plan: run `sourcecred score` in an existing instance. Scores should
not change. Run `sourcecred score` in an instance with unset ids in the
dependency file, and verify that they are set and serialized to both the
ledger and the dependencies file correctly. (nb. the template instance
is fine for this purpose.)