add IntervalSequence
An IntervalSequence is an opaque read-only array of intervals, with
runtime validation that the intervals are non-overlapping and also have
no gaps between them, and that every interval has non-negative duration.
This way we can use intervals and not worry that the interval may be
out-of-order or have gaps, etc.

Test plan: Unit tests added to verify that the interval sequences have
the expected properties. Also, `git grep -F "$ReadOnlyArray<Interval>"`
only has hits within the interval module itself.