github: decouple error types from dispositions (#2326)
Summary:
Previously, our GitHub fetch errors included a `retry` flag in the error
itself (with a value mandated by the variant). We now care about more
than just a boolean “retry” flag, since the correct response to an error
may be to wait until a certain time. Instead of just changing the field
on the error values, we introduce a new `errorDisposition` to specify
how to respond to a given error. This opens the door to more substantial
logic, like querying for the time at which quotas refresh, or retrying
only some kinds of `FETCH_ERROR`s (`ETIMEDOUT`).

wchargin-branch: github-error-decouple-retry