github: request quota refresh time (#2327)
Summary:
Upon exhausting its quota, the GitHub plugin now asks GitHub when the
quota will be refreshed and waits until that time, rather than just
blindly waiting 15 minutes repeatedly. (You can query for the rate limit
information even if your quota is exhausted.) The new logic is gated
behind a blanket `try`-`catch`; if anything goes wrong, we fall back to
waiting for 15 minutes.

As an incidental change in this PR, we remove the dependency injection
of `fetch`. It was just kind of confusing, and wasn’t helpful; this
module’s networking code has no unit tests that would benefit from the
injection.

We also now log to the console while waiting for a refresh, which seems
nicer than hanging for up to an hour with no output.

Test Plan:
Unit tests included for parts of the logic. As an end-to-end test, I ran
four concurrent loads of [the `1Hive/pollen` instance][1]. This
exhausted my quota in about 10 minutes, after which it properly waited
for refresh and then resumed:

```
  GO   sourcecred/github: github: loading 1Hive/gardens-template
 DONE  sourcecred/github: github: loading 1Hive/gardens-template: 9123ms
  GO   sourcecred/github: github: loading 1Hive/honey-template
 DONE  sourcecred/github: github: loading 1Hive/honey-template: 5766ms
  GO   sourcecred/github: github: loading 1Hive/transactions-app
rate limit exceeded; waiting for refresh at 2020-10-04T21:57:24.000Z
 DONE  sourcecred/github: github: loading 1Hive/transactions-app: 37m 2s
  GO   sourcecred/github: github: loading 1Hive/flora
 DONE  sourcecred/github: github: loading 1Hive/flora: 8714ms
  GO   sourcecred/github: github: loading 1Hive/apiary
 DONE  sourcecred/github: github: loading 1Hive/apiary: 44s
[...]
 DONE  loading sourcecred/github: 1h 9m
 DONE  load: 1h 9m
```

(Note the 37-minute elapsed time for `1Hive/transactions-app`.) The
target time was indeed 60 seconds after my token would be refreshed:

```
$ ghquery '{ rateLimit { remaining resetAt } }'
{"data":{"rateLimit":{"remaining":0,"resetAt":"2020-10-04T21:56:24Z"}}}
```

[1]: https://github.com/1Hive/pollen/tree/55a2a57ee3f8969189af3964865cff0262778653

wchargin-branch: github-ask-for-refresh-time